# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Space DOGE API",
    "version": "1.0.0",
    "description": "Wallet-first Coin Rush Arena with verifiable Scrypt cloud-mining rewards. Every endpoint below reflects the actual current implementation \u2014 request/response shapes were transcribed directly from the route handlers, not hand-guessed.\n\n**Auth**: almost every endpoint requires a session, established via `/api/auth/verify` (SIWE sign-in), which sets an httpOnly `SpaceDOGE_session` cookie. Admin endpoints (`/api/admin/*`) additionally require the session's wallet address to be on the admin allowlist (env `ADMIN_ADDRESSES` or the `AdminUser` table) \u2014 a non-admin session hitting those returns `403`, not `401`.\n\n**Money fields**: all Decimal/Prisma fields are converted to plain JSON `number` before being returned \u2014 never strings.\n\n**Timestamps**: all dates are ISO-8601 strings in responses.",
    "contact": {
      "name": "Space DOGE"
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "Same-origin"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "SpaceDOGE_session",
        "description": "httpOnly JWT session cookie set by POST /api/auth/verify."
      },
      "cronBearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "Static shared secret (env DEPOSIT_CRON_SECRET), used only by the deposit-watch cron trigger."
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ]
      },
...
      "WithdrawChainConfig": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "chainKey": {
            "type": "string",
            "example": "DOGE"
          },
          "label": {
            "type": "string"
          },
          "coinSymbol": {
            "type": "string",
            "enum": [
              "USDT",
              "DOGE"
            ]
          },
          "addressRegex": {
            "type": "string"
          },
          "explorerTxUrl": {
            "type": "string",
            "nullable": true
          },
          "enabled": {
            "type": "boolean"
          },
          "sortOrder": {
            "type": "integer"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    },
    "responses": {
      "NotAuthenticated": {
        "description": "No valid session cookie",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "error": "Not authenticated"
            }
          }
        }
      },
      "NotAuthorized": {
        "description": "Session valid but wallet is not an admin",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "error": "Not authorized"
            }
          }
        }
      },
      "BadRequest": {
        "description": "Validation failure",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    }
  },
  "security": [
    {
      "sessionCookie": []
    }
  ],
  "paths": {
...
    "/api/auth/onboard": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Record age confirmation, country, and terms acceptance",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "countryCode",
                  "ageConfirmed"
                ],
                "properties": {
                  "countryCode": {
                    "type": "string",
                    "minLength": 2,
                    "maxLength": 2,
                    "example": "US"
                  },
                  "ageConfirmed": {
                    "type": "boolean",
                    "enum": [
                      true
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    },
                    "termsVersion": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/NotAuthenticated"
          }
        }
      }
    },
...
    "/api/admin/overview": {
      "get": {
        "tags": [
          "Admin: Overview"
        ],
        "summary": "Admin dashboard aggregate stats over a configurable day range",
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "parameters": [
          {
            "name": "days",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 7,
              "maximum": 90,
              "default": 30
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK \u2014 wallet counts, balances, deposit/withdrawal totals + daily series, mining epoch summary",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "range": {
                      "type": "integer"
                    },
                    "walletCount": {
                      "type": "integer"
                    },
                    "newWallets": {
                      "type": "integer"
                    },
                    "matchCount": {
                      "type": "integer"
                    },
                    "balances": {
                      "type": "object",
                      "description": "Raw ledger sums by balance type, across all real wallets (excludes bot: and platform: pseudo-wallets except treasuryUsdt, which IS the platform:treasury wallet's own balance).",
                      "properties": {
                        "playUsdt": {
                          "type": "number"
                        },
                        "gameRewardUsdt": {
                          "type": "number"
                        },
                        "pts": {
                          "type": "number"
                        },
                        "pendingDoge": {
                          "type": "number"
                        },
                        "availableDoge": {
                          "type": "number"
                        },
                        "recycledUsdt": {
                          "type": "number"
                        },
                        "referralUsdt": {
                          "type": "number"
                        },
                        "treasuryUsdt": {
                          "type": "number",
                          "description": "The platform treasury's real balance \u2014 30% platform fee net of referral commission paid out, credited at match entry. See src/lib/treasury.ts."
                        }
                      }
                    },
                    "availableUsdt": {
                      "type": "number",
                      "description": "User-withdrawable USDT liability \u2014 sum of RECYCLED_USDT + REFERRAL_USDT (the two balance types /api/wallet/withdraw accepts)."
                    },
                    "deposits": {
                      "type": "object",
                      "description": "Totals by status, a daily series, and the 8 most recent rows."
                    },
                    "withdrawals": {
                      "type": "object",
                      "description": "Totals by status, fees collected, a daily series, and the 8 most recent rows."
                    },
                    "mining": {
                      "type": "object",
                      "description": "Active mining power/contract counts and the latest epoch summary."
                    }
                  }
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/NotAuthorized"
          }
        }
      }
    },
...
    "/api/admin/settings/withdraw-chains/{id}": {
      "patch": {
        "tags": [
          "Admin: Settings"
        ],
        "summary": "Edit or toggle an existing withdrawal chain (chainKey/coinSymbol immutable)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "label": {
                    "type": "string"
                  },
                  "addressRegex": {
                    "type": "string"
                  },
                  "explorerTxUrl": {
                    "type": "string",
                    "nullable": true
                  },
                  "enabled": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "row": {
                      "$ref": "#/components/schemas/WithdrawChainConfig"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid regex or bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/NotAuthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "delete": {
        "tags": [
          "Admin: Settings"
        ],
        "summary": "Remove a withdrawal chain option",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/NotAuthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    }
  }
}